feat: preauthorize with #index? on connection types
The `preauthorize: { <...> }` field extension automatically determines
the `to:` parameter based on the field type if not provided. Previously
it used `index?` in case the field's type is a list and `show?`
otherwise.

An alternative to lists are relay connection. Action Policy GraphQL
handles them now similar to lists and by default uses `index?` instead
of the default `show?`.